[Telemetry] Terminate browsers more aggressively.
This is safe to do when we're sure the profile will never be reused
and speeds up telemetry_unittests by about 7%.

BUG=388256

Review URL: https://codereview.chromium.org/476903002

Cr-Commit-Position: refs/heads/master@{#289720}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289720 0039d316-1c4b-4281-b951-d872f2087c98

+++ telemetry/telemetry/core/backends/chrome/desktop_browser_backend.py
@@ -317,8 +317,8 @@ class DesktopBrowserBackend(chrome_browser_backend.ChromeBrowserBackend):
   def Close(self):
     super(DesktopBrowserBackend, self).Close()
 
-    # First, try to politely shutdown.
-    if self.IsBrowserRunning():
+    # Shutdown politely if the profile may be used again.
+    if self._output_profile_path and self.IsBrowserRunning():
       self._proc.terminate()
       try:
         util.WaitFor(lambda: not self.IsBrowserRunning(), timeout=5)
@@ -326,7 +326,7 @@ class DesktopBrowserBackend(chrome_browser_backend.ChromeBrowserBackend):
       except util.TimeoutException:
         logging.warning('Failed to gracefully shutdown. Proceeding to kill.')
 
-    # If it didn't comply, get more aggressive.
+    # Shutdown aggressively if the above failed or if the profile is temporary.
     if self.IsBrowserRunning():
       self._proc.kill()
